Port live-reload into esbuild `watch` server
Instead of running a customer websocket server, I'm using esbuilds watch
server and subscribing to it's `EventSource` stream to trigger the
reload for the renderer and sandbox threads. This makes more sense than
the sockets, since the "reload" message only needs to be
uni-directional.

`ui.html` still has to be copied over to `dist` manually